Move balance tests to test_utilities.py
